bots/Wishlist_bot/requirements.txt
discord.py>=2.3.2
notion-client>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
python-dateutil>=2.8.2
pytz>=2022.7
beautifulsoup4>=4.12.0
lxml>=4.9.2
aiohttp>=3.8.4
Pillow>=9.5.0
openai>=1.0.0 